Conflicts:
	src/gamecontent/MyController.java

+++ src/game/gameobject/Stats.java
@@ -1,5 +1,10 @@
 package game.gameobject;
 
+
+import collision.interactive.InteractiveResponse;
+
+import static collision.interactive.InteractiveResponse.*;
+
 /**
  * Created by przemek on 10.08.15.
  */
@@ -9,18 +14,43 @@ public abstract class Stats {
     protected int health = 100;
     protected int maxHealth = 100;
     protected int strength = 2;
+    protected int defence = 2;
+    protected float sideDefenceModifier = 10;
+    protected float backDefenceModifier = 4;
+    protected float protection = 50;
+    protected float protectionSideModifier = 4;
+    protected float protectionBackModifier = 1;
+    protected boolean protectionState;
 
 
     public Stats(GameObject owner) {
         this.owner = owner;
     }
 
-    public void decreaseHealth(int value) {
-        health -= value;
-        if (health <= 0) {
-            owner.delete();
-            System.out.println(owner.getName() + " zginał.");
+    // TODO ustawienie zachowania w zależności od potwora/gracza
+
+    public void decreaseHealth(InteractiveResponse response) {
+        if (health > 0) {
+            int hurt = 0;
+            switch (response.getDirection()) {
+                case FRONT:
+                    hurt = Math.round(response.getData() / (defence * (protectionState ? protection : 1)));
+                    break;
+                case BACK:
+                    hurt = Math.round(response.getData() / (defence * (protectionState ? protection * protectionBackModifier : backDefenceModifier)));
+                    break;
+                case SIDE:
+                    hurt = Math.round(response.getData() / (defence * (protectionState ? protection * protectionSideModifier : sideDefenceModifier)));
+                    break;
+            }
+            health -= hurt;
+            System.out.println(owner.getName() + " dostał za " + hurt);
+            if (health <= 0) {
+                owner.delete();
+                System.out.println(owner.getName() + " zginał.");
+            }
         }
+
     }
 
     public int getHealth() {
@@ -41,5 +71,13 @@ public abstract class Stats {
 
     public void setStrength(int strength) {
         this.strength = strength;
+    }
+
+    public boolean isProtectionState() {
+        return protectionState;
+    }
+
+    public void setProtectionState(boolean protectionState) {
+        this.protectionState = protectionState;
     }
 }
